src: use non-deprecated overload of V8::SetFlagsFromString

+++ src/node.cc
@@ -939,7 +939,8 @@ void Init(int* argc,
   }
 
   if (per_process::cli_options->print_v8_help) {
-    V8::SetFlagsFromString("--help", 6);  // Doesn't return.
+    // Doesn't return.
+    V8::SetFlagsFromString("--help", static_cast<size_t>(6));
     UNREACHABLE();
   }
 
@@ -997,7 +998,8 @@ InitializationResult InitializeOncePerProcess(int argc, char** argv) {
   }
 
   if (per_process::cli_options->print_v8_help) {
-    V8::SetFlagsFromString("--help", 6);  // Doesn't return.
+    // Doesn't return.
+    V8::SetFlagsFromString("--help", static_cast<size_t>(6));
     UNREACHABLE();
   }
